source_code/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-RESTX==1.1.0
pymongo==4.5.0
pika==1.3.1
python-dotenv==1.0.0
mysql-connector-python==8.2.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
PyJWT==2.8.0
gunicorn==21.2.0
pytest==7.4.2
pytest-cov==4.1.0 